--- a/scripts/git-commit-id.sh
+++ b/scripts/git-commit-id.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-echo "const CURRENTCOMMIT* = \"$(git rev-parse HEAD)"\"\ > src/catnaplib/global/currentcommit.nim
+echo "const CURRENTCOMMIT* = static: \"$(git rev-parse HEAD)"\"\ > src/catnaplib/global/currentcommit.nim